fix: Handle vendor search errors
