+++ apps/console/src/components/pages/protected/contacts/create/form/fields/vendor-search-state.test.ts
@@ -1,5 +1,5 @@
 import { keepPreviousData, QueryClient, QueryObserver } from '@tanstack/react-query'
-import { getVendorSearchState } from './vendor-select-field'
+import { getVendorSearchState } from './vendor-search-state'
 
 describe('getVendorSearchState', () => {
   it('is idle and hides results when a cleared search leaves placeholder data on a disabled query', async () => {
@@ -29,19 +29,24 @@ describe('getVendorSearchState', () => {
         isSearchSettled: true,
         isLoading: result.isLoading,
         isPlaceholderData: result.isPlaceholderData,
+        isError: result.isError,
       }),
     ).toEqual({ canShowResults: false, isPending: false })
   })
 
   it.each([
-    { isSearchSettled: false, isLoading: false, isPlaceholderData: false },
-    { isSearchSettled: true, isLoading: true, isPlaceholderData: false },
-    { isSearchSettled: true, isLoading: false, isPlaceholderData: true },
+    { isSearchSettled: false, isLoading: false, isPlaceholderData: false, isError: false },
+    { isSearchSettled: true, isLoading: true, isPlaceholderData: false, isError: false },
+    { isSearchSettled: true, isLoading: false, isPlaceholderData: true, isError: false },
   ])('remains pending while an active search is not ready', (queryState) => {
     expect(getVendorSearchState({ searchText: 'acme', ...queryState })).toEqual({ canShowResults: false, isPending: true })
   })
 
   it('shows results when the active search is ready', () => {
-    expect(getVendorSearchState({ searchText: 'acme', isSearchSettled: true, isLoading: false, isPlaceholderData: false })).toEqual({ canShowResults: true, isPending: false })
+    expect(getVendorSearchState({ searchText: 'acme', isSearchSettled: true, isLoading: false, isPlaceholderData: false, isError: false })).toEqual({ canShowResults: true, isPending: false })
+  })
+
+  it('stops pending and hides results when the search fails', () => {
+    expect(getVendorSearchState({ searchText: 'acme', isSearchSettled: true, isLoading: false, isPlaceholderData: false, isError: true })).toEqual({ canShowResults: false, isPending: false })
   })
 })
